feature/balance refactor - add new packages
