+++ src/main/java/com/isuru/bank/customer/Customer.java
@@ -1,0 +1,10 @@
+package com.isuru.bank.customer;
+
+public interface Customer {
+
+    void setAmount(String date, double amount, boolean isDebit);
+
+    double getMonthlyBalance(String month);
+
+    double getCurrentAmount();
+}
